+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 * **New**
     * Added `dusty doctor` command to attempt to fix some common VirtualBox networking problems
+    * Added a custom 502 page for apps which use host forwarding. This lets you see the status and logs of your container while you're waiting for its HTTP service to come online
 
 ## 0.7.0 (February 18, 2016)